fix(playback): suppress error dialog for recoverable MediaCodec state errors
MusicService.onPlayerError already auto-recovers MediaCodec decoder-state
faults (ALAC c2.mtk.alac.decoder 'Error 0x80000000', queueInputBuffer
Released-state races, decoder init failures) by re-preparing the player.
The song resumes playback automatically after recovery succeeds.

However, PlayerConnection.onPlayerErrorChanged fires immediately when the
error surfaces — before recovery completes — so the PlaybackErrorDialog
briefly flashes on screen, then auto-dismisses once recovery succeeds.
The user perceives this as a spurious error popup.

Fix: extract the codec-error classifier into a shared
isRecoverableMediaCodecStateError() util (single source of truth — same
logic MusicService uses to decide whether to retry). Have PlayerConnection
suppress these recoverable errors from the UI error flow so the dialog
never shows for them. The dialog still appears for genuinely unrecoverable
errors (login required, no internet, HTTP failures, exhausted retry budget,
etc.).

MusicService.isMediaCodecStateError is refactored to delegate to the new
util — pure refactor, no behavior change to the recovery logic.
